Implement batch queries for MSSQL Client
Closes #607
